Playback v3 (#43)
Moved ContextExplorer to AMAK, it's now the VuiExplorer. Provide the same functionnality as the ContextExplorer, but for all VUIs !

Modification to Drawables, allowing default usage with the VuiExplorer, and ease of customization.

Updated AMAK's exemples to use VuiExplorer.

Updated ContextRendererFX to make use of VuiExplorer. Allowing huges simplifications of the codebase.

BackupSystem now save the local model and other context info, allowing for a more detailed preview in the save explorer.

Added a dimension selector in save explorer.

'Request Here' and 'Learn Here' (right-click on the main vui) now work in any dimension.

DrawablePoint now use a + icon.

Add a 'quick save' button in the main toolbar.